\name{fertilizer}
\alias{fertilizer}
\docType{data}
\title{ Fertilizer data set }

\description{ 
the fertilizer dataset contains \eqn{4} features and \eqn{96} records. Results from an experiment to compare yields of a crop obtained under three different fertilizers. the target feature is \emph{yield}.
}

\usage{ data(fertilizer) }

\seealso{ \code{\link[liver]{bank}}, 
          \code{\link[liver]{churn_mlc}}, 
          \code{\link[liver]{churn}}, 
          \code{\link[liver]{churn_tel}}, 
          \code{\link[liver]{adult}}, 
          \code{\link[liver]{risk}}, 
          \code{\link[liver]{cereal}}, 
          \code{\link[liver]{advertising}}, 
          \code{\link[liver]{marketing}}, 
          \code{\link[liver]{drug}}, 
          \code{\link[liver]{house}}, 
          \code{\link[liver]{house_price}}, 
          \code{\link[liver]{red_wines}}, 
          \code{\link[liver]{white_wines}}, 
          \code{\link[liver]{insurance}}, 
          \code{\link[liver]{caravan}}, 
          \code{\link[liver]{corona}}
          }

\examples{
data(fertilizer)
str(fertilizer)
}

\keyword{ datasets }
