\name{housePrice}
\alias{housePrice}

\docType{data}

\title{ housePrice dataset }

\description{ 
This data set contains \eqn{1460} rows and \eqn{81} columns (features). The "SalePrice" column is the target.
}

\usage{ data( housePrice ) }

\format{ 
The \code{housePrice} dataset, as a data frame, contains \eqn{1460} rows and \eqn{81} columns (variables/features). 
}

\details{
For more information related to the dataset see:

\url{https://www.kaggle.com/c/house-prices-advanced-regression-techniques/data}
}

\seealso{ \code{\link{adult}}, \code{\link{classifyRisk}}, \code{\link{churn}}, \code{\link{churnTel}}, \code{\link{bank}}, \code{\link{advertising}}, \code{\link{marketing}}, \code{\link{insurance}}, \code{\link{cereal}}, \code{\link{house}} }

\examples{
data( housePrice )

str( housePrice )
}

\keyword{ datasets }
