\name{liver-package}
\alias{liver-package}

\docType{package}

\title{ liver: "Eating the Liver of Data Science" }

\description{
The \pkg{liver} package provides a collection of helper functions that make various techniques from data science more user-friendly for non-experts. In this way, our aim is to allow non-experts to become familiar with the techniques with only a minimal level of coding knowledge. Indeed, following an ancient Persian idiom, we refer to this as "eating the liver of data science" which could be interpreted as "getting intimately close with data science". Examples of procedures we include are: data partitioning for out-of-sample testing, computing Mean Squared Error (MSE) for quantifying prediction accuracy, and data transformation (z-score and min-max). Besides such helper functions, the package also includes several interesting datasets that are useful for multivariate analysis.
}

\author{ 

	Reza Mohammadi \email{a.mohammadi@uva.nl} \cr
	Amsterdam Business School \cr
	University of Amsterdam \cr
	
	Kevin Burke \email{kevin.burke@ul.ie} \cr
	Departement of Statistics \cr
	University of Limerick \cr

	Maintainer: Reza Mohammadi \email{a.mohammadi@uva.nl}
}

\keyword{package}
