\name{fertilizer}
\alias{fertilizer}
\docType{data}
\title{ Fertilizer data set }

\description{ 
The fertilizer dataset contains \eqn{4} features and \eqn{96} records. Results from an experiment to compare yields of a crop obtained under three different fertilizers. The target feature is \emph{yield}.
}

\usage{ data( fertilizer ) }

\seealso{ \code{\link{adult}}, \code{\link{risk}}, \code{\link{churn}}, \code{\link{churnTel}}, \code{\link{bank}}, \code{\link{advertising}}, \code{\link{marketing}}, \code{\link{insurance}}, \code{\link{cereal}}, \code{\link{housePrice}}, \code{\link{house}} }

\examples{
data( fertilizer )

str( fertilizer )
}

\keyword{ datasets }
