\name{fertilizer}
\alias{fertilizer}
\docType{data}
\title{ Fertilizer data set }

\description{ 
the fertilizer dataset contains \eqn{4} features and \eqn{96} records. Results from an experiment to compare yields of a crop obtained under three different fertilizers. the target feature is \emph{yield}.
}

\usage{ data( fertilizer ) }

\seealso{ \code{\link[liver]{adult}}, \code{\link[liver]{risk}}, \code{\link[liver]{churn}}, \code{\link[liver]{churnTel}}, \code{\link[liver]{bank}}, \code{\link[liver]{advertising}}, \code{\link[liver]{marketing}}, \code{\link[liver]{insurance}}, \code{\link[liver]{cereal}}, \code{\link[liver]{housePrice}}, \code{\link[liver]{house}} }

\examples{
data( fertilizer )

str( fertilizer )
}

\keyword{ datasets }
