\name{liver-package}
\alias{liver-package}

\docType{package}

\title{ liver: "Eating the Liver of Data Science" }

\description{
The \pkg{liver} package provides a suite of helper functions and a collection of datasets used in the book \href{https://uncovering-data-science.netlify.app}{Data Science Foundations and Machine Learning Using R}. Designed to make data science techniques accessible to individuals with minimal coding experience, it simplifies tasks such as data partitioning for out-of-sample testing and data transformations (z-score and min-max). Inspired by an ancient Persian idiom, the package likens this learning process to "eating the liver of data science," symbolizing deep and immersive engagement with the field. In addition to its helper functions, liver also includes a collection of datasets valuable for multivariate analysis.
}

\author{ 

	Reza Mohammadi \email{a.mohammadi@uva.nl} \cr
	Amsterdam Business School \cr
	University of Amsterdam \cr
	
	Kevin Burke \email{kevin.burke@ul.ie} \cr
	Departement of Statistics \cr
	University of Limerick \cr

	Maintainer: Reza Mohammadi \email{a.mohammadi@uva.nl}
}

\keyword{package}
