\name{liver-package}
\alias{liver-package}

\docType{package}

\title{ liver: Foundations Toolkit and Datasets for Data Science }

\description{

The \pkg{liver} package provides a collection of helper functions and illustrative datasets to support learning and teaching of data science with R. The package is designed as a companion to the book \href{https://book-data-science-r.netlify.app}{Data Science Foundations and Machine Learning Using R}, making key data science techniques accessible to individuals with minimal coding experience. Functions include tools for data partitioning, performance evaluation, and data transformations (e.g., z-score and min-max scaling). The included datasets are curated to highlight practical applications in data exploration, modeling, and multivariate analysis. An early inspiration for the package came from an ancient Persian idiom about "eating the liver," symbolizing deep and immersive engagement with knowledge.
}

\author{ 

	Reza Mohammadi \email{a.mohammadi@uva.nl} \cr
	Amsterdam Business School \cr
	University of Amsterdam \cr
	
	Kevin Burke \email{kevin.burke@ul.ie} \cr
	Departement of Statistics \cr
	University of Limerick \cr

	Maintainer: Reza Mohammadi \email{a.mohammadi@uva.nl}
}

\keyword{package}
