\name{liver-package}
\alias{liver-package}

\docType{package}

\title{ liver: "Eating the Liver of Data Science" }

\description{

The \pkg{liver} package offers a suite of helper functions to simplify various data science techniques for non-experts. This package aims to enable individuals with only a minimal level of coding knowledge to become acquainted with these techniques in an accessible manner. Inspired by an ancient Persian idiom, we liken this process to "eating the liver of data science," suggesting a deep and intimate engagement with the field of data science. This package includes functions for tasks such as data partitioning for out-of-sample testing, calculating Mean Squared Error (MSE) to assess prediction accuracy, and data transformations (z-score and min-max). In addition to these helper functions, the 'liver' package also features several intriguing datasets valuable for multivariate analysis.
}

\author{ 

	Reza Mohammadi \email{a.mohammadi@uva.nl} \cr
	Amsterdam Business School \cr
	University of Amsterdam \cr
	
	Kevin Burke \email{kevin.burke@ul.ie} \cr
	Departement of Statistics \cr
	University of Limerick \cr

	Maintainer: Reza Mohammadi \email{a.mohammadi@uva.nl}
}

\keyword{package}
